logs page wrapper
